feat: restructure and split off applications codebase into ibc-apps dir
