+++ crates/ibc-apps/transfer/src/handler/send_transfer.rs
@@ -1,18 +1,15 @@
-use crate::applications::transfer::context::{
-    TokenTransferExecutionContext, TokenTransferValidationContext,
-};
-use crate::applications::transfer::error::TokenTransferError;
-use crate::applications::transfer::events::TransferEvent;
-use crate::applications::transfer::msgs::transfer::MsgTransfer;
-use crate::applications::transfer::{is_sender_chain_source, MODULE_ID_STR};
-use crate::core::events::{MessageEvent, ModuleEvent};
-use crate::core::ics04_channel::context::{
-    SendPacketExecutionContext, SendPacketValidationContext,
-};
-use crate::core::ics04_channel::handler::send_packet::{send_packet_execute, send_packet_validate};
-use crate::core::ics04_channel::packet::Packet;
-use crate::core::ics24_host::path::{ChannelEndPath, SeqSendPath};
-use crate::prelude::*;
+use ibc::core::events::{MessageEvent, ModuleEvent};
+use ibc::core::ics04_channel::context::{SendPacketExecutionContext, SendPacketValidationContext};
+use ibc::core::ics04_channel::handler::send_packet::{send_packet_execute, send_packet_validate};
+use ibc::core::ics04_channel::packet::Packet;
+use ibc::core::ics24_host::path::{ChannelEndPath, SeqSendPath};
+use ibc::prelude::*;
+use ibc_app_transfer_types::error::TokenTransferError;
+use ibc_app_transfer_types::events::TransferEvent;
+use ibc_app_transfer_types::msgs::transfer::MsgTransfer;
+use ibc_app_transfer_types::{is_sender_chain_source, MODULE_ID_STR};
+
+use crate::context::{TokenTransferExecutionContext, TokenTransferValidationContext};
 
 /// Initiate a token transfer. Equivalent to calling [`send_transfer_validate`], followed by [`send_transfer_execute`].
 pub fn send_transfer<SendPacketCtx, TokenCtx>(
