+++ crates/ibc-apps/transfer/types/src/amount.rs
@@ -4,6 +4,7 @@ use core::ops::Deref;
 use core::str::FromStr;
 
 use derive_more::{Display, From, Into};
+use ibc::prelude::*;
 use primitive_types::U256;
 
 use super::error::TokenTransferError;
@@ -11,7 +12,6 @@ use super::error::TokenTransferError;
 use crate::alloc::borrow::ToOwned;
 #[cfg(feature = "schema")]
 use crate::alloc::string::String;
-use crate::prelude::*;
 
 /// A type for representing token transfer amounts.
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
@@ -19,7 +19,7 @@ use crate::prelude::*;
 #[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord, Display, From, Into)]
 pub struct Amount(
     #[cfg_attr(feature = "schema", schemars(with = "String"))]
-    #[serde(serialize_with = "crate::serializers::serde_string::serialize")]
+    #[serde(serialize_with = "ibc::serializers::serde_string::serialize")]
     #[serde(deserialize_with = "deserialize")]
     U256,
 );
