+++ crates/ibc-apps/transfer/types/src/error.rs
@@ -4,13 +4,12 @@ use core::convert::Infallible;
 use core::str::Utf8Error;
 
 use displaydoc::Display;
+use ibc::core::ics04_channel::acknowledgement::StatusValue;
+use ibc::core::ics04_channel::channel::Order;
+use ibc::core::ics24_host::identifier::{ChannelId, IdentifierError, PortId};
+use ibc::core::ContextError;
+use ibc::prelude::*;
 use uint::FromDecStrErr;
-
-use crate::core::ics04_channel::acknowledgement::StatusValue;
-use crate::core::ics04_channel::channel::Order;
-use crate::core::ics24_host::identifier::{ChannelId, IdentifierError, PortId};
-use crate::core::ContextError;
-use crate::prelude::*;
 
 #[derive(Display, Debug)]
 pub enum TokenTransferError {
