+++ crates/ibc-apps/transfer/types/src/packet.rs
@@ -4,13 +4,13 @@ use alloc::string::ToString;
 use core::convert::TryFrom;
 use core::str::FromStr;
 
+use ibc::Signer;
 use ibc_proto::ibc::applications::transfer::v2::FungibleTokenPacketData as RawPacketData;
 
 use super::error::TokenTransferError;
 use super::{Amount, Memo, PrefixedCoin, PrefixedDenom};
 #[cfg(feature = "schema")]
 use crate::alloc::borrow::ToOwned;
-use crate::signer::Signer;
 
 /// Defines the structure of token transfers' packet bytes
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
@@ -69,7 +69,7 @@ mod tests {
     use primitive_types::U256;
 
     use super::*;
-    use crate::applications::transfer::BaseCoin;
+    use crate::BaseCoin;
 
     impl PacketData {
         pub fn new_dummy() -> Self {
