+++ crates/ibc-testkit/src/testapp/ibc/applications/transfer/context.rs
@@ -1,11 +1,11 @@
-use ibc::applications::transfer::context::{
-    cosmos_adr028_escrow_address, TokenTransferExecutionContext, TokenTransferValidationContext,
-};
-use ibc::applications::transfer::error::TokenTransferError;
-use ibc::applications::transfer::PrefixedCoin;
 use ibc::core::ics24_host::identifier::{ChannelId, PortId};
 use ibc::prelude::*;
 use ibc::Signer;
+use ibc_app_transfer::context::{
+    cosmos_adr028_escrow_address, TokenTransferExecutionContext, TokenTransferValidationContext,
+};
+use ibc_app_transfer::types::error::TokenTransferError;
+use ibc_app_transfer::types::PrefixedCoin;
 use subtle_encoding::bech32;
 
 use super::types::DummyTransferModule;
