--- crates/ibc/src/applications/mod.rs
@@ -1,4 +1,0 @@
-//! Implementation of IBC applications
-
-#[cfg(feature = "serde")]
-pub mod transfer;
