+++ crates/ibc/src/core/ics04_channel/acknowledgement.rs
@@ -140,71 +140,3 @@ impl From<AcknowledgementStatus> for Acknowledgement {
             .expect("token transfer internal error: ack is never supposed to be empty")
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::applications::transfer::ack_success_b64;
-    use crate::applications::transfer::error::TokenTransferError;
-
-    #[test]
-    fn test_ack_ser() {
-        fn ser_json_assert_eq(ack: AcknowledgementStatus, json_str: &str) {
-            let ser = serde_json::to_string(&ack).unwrap();
-            assert_eq!(ser, json_str)
-        }
-
-        ser_json_assert_eq(
-            AcknowledgementStatus::success(ack_success_b64()),
-            r#"{"result":"AQ=="}"#,
-        );
-        ser_json_assert_eq(
-            AcknowledgementStatus::error(TokenTransferError::PacketDataDeserialization.into()),
-            r#"{"error":"failed to deserialize packet data"}"#,
-        );
-    }
-
-    #[test]
-    fn test_ack_success_to_vec() {
-        let ack_success: Vec<u8> = AcknowledgementStatus::success(ack_success_b64()).into();
-
-        // Check that it's the same output as ibc-go
-        // Note: this also implicitly checks that the ack bytes are non-empty,
-        // which would make the conversion to `Acknowledgement` panic
-        assert_eq!(ack_success, r#"{"result":"AQ=="}"#.as_bytes());
-    }
-
-    #[test]
-    fn test_ack_error_to_vec() {
-        let ack_error: Vec<u8> =
-            AcknowledgementStatus::error(TokenTransferError::PacketDataDeserialization.into())
-                .into();
-
-        // Check that it's the same output as ibc-go
-        // Note: this also implicitly checks that the ack bytes are non-empty,
-        // which would make the conversion to `Acknowledgement` panic
-        assert_eq!(
-            ack_error,
-            r#"{"error":"failed to deserialize packet data"}"#.as_bytes()
-        );
-    }
-
-    #[test]
-    fn test_ack_de() {
-        fn de_json_assert_eq(json_str: &str, ack: AcknowledgementStatus) {
-            let de = serde_json::from_str::<AcknowledgementStatus>(json_str).unwrap();
-            assert_eq!(de, ack)
-        }
-
-        de_json_assert_eq(
-            r#"{"result":"AQ=="}"#,
-            AcknowledgementStatus::success(ack_success_b64()),
-        );
-        de_json_assert_eq(
-            r#"{"error":"failed to deserialize packet data"}"#,
-            AcknowledgementStatus::error(TokenTransferError::PacketDataDeserialization.into()),
-        );
-
-        assert!(serde_json::from_str::<AcknowledgementStatus>(r#"{"success":"AQ=="}"#).is_err());
-    }
-}
