+++ crates/ibc/src/core/ics04_channel/handler.rs
@@ -1,13 +1,13 @@
 //! This module implements the processing logic for ICS4 (channel) messages.
 
-pub(crate) mod acknowledgement;
-pub(crate) mod chan_close_confirm;
-pub(crate) mod chan_close_init;
-pub(crate) mod chan_open_ack;
-pub(crate) mod chan_open_confirm;
-pub(crate) mod chan_open_init;
-pub(crate) mod chan_open_try;
-pub(crate) mod recv_packet;
-pub(crate) mod send_packet;
-pub(crate) mod timeout;
-pub(crate) mod timeout_on_close;
+pub mod acknowledgement;
+pub mod chan_close_confirm;
+pub mod chan_close_init;
+pub mod chan_open_ack;
+pub mod chan_open_confirm;
+pub mod chan_open_init;
+pub mod chan_open_try;
+pub mod recv_packet;
+pub mod send_packet;
+pub mod timeout;
+pub mod timeout_on_close;
